Get All favorites OK Restructurations pour maintenabilité
