+++ src/app/pages/favorites/favorites.component.html
@@ -1,0 +1,1 @@
+<p>favorites works!</p>
