+++ src/app/pages/login/login.component.scss
@@ -32,6 +32,7 @@ h3 {
 
 mat-form-field {
   width: 100%;
+  
 }
 
 mat-label {
@@ -40,7 +41,8 @@ mat-label {
 }
 
 input[matInput] {
-  border: 1px solid #ddd;
+  //border: 1px solid #ddd;
+  width: 94%;
   padding: 10px;
   border-radius: 5px;
   transition: border-color 0.3s ease, box-shadow 0.3s ease;
