+++ src/app/pages/register/register.component.scss
@@ -40,7 +40,8 @@ mat-label {
 }
 
 input[matInput] {
-  border: 1px solid #ddd;
+  //border: 1px solid #ddd;
+  width: 94%;
   padding: 10px;
   border-radius: 5px;
   transition: border-color 0.3s ease, box-shadow 0.3s ease;
